[feat] 볼링게임 엔진 분리
