+++ src/main/java/com/jaeyeonling/bowling/view/console/ConsoleOutputView.java
@@ -9,6 +9,7 @@ public final class ConsoleOutputView {
     private static final String BOWLING_GAME_HEADER =
             "|  NAME |  01   |  02   |  03   |  04   |  05   |  06   |  07   |  08   |  09   |  10   |";
     private static final String RESULT_MESSAGE = "게임이 끝났습니다.";
+
     private static final PrintStream CONSOLE = System.out;
 
     private ConsoleOutputView() { }
